lnwallet: fully validate all channel constraints in validateCommitmentSanity
This commit introduces changes to the validateCommitmentSanity
function to fully validate all channel constraints.
validateCommitmentSanity now validates that the
MaxPendingAmount, ChanReserve, MinHTLC, & MaxAcceptedHTLCs
limits are all adhered to during the lifetime of a channel.

evaluateHTLCView is made optionally non-mutable, such that
it can be used to calculate the resulting balances from
applying the given updates. Then the channel constraints
are validated from the point-of-view of the local and the
remote node, to make sure both will accept the updates.

Unit tests were also included to account for validateCommitmentSanity
changes.

Co-authored-by: nsa <elzeigel@gmail.com>