chainntnfs/btcdnotify: only notify spend on confirmation
This commit changes the btcd notifier to only notify about a spend on
confirmation of the spending transaction. This is done to account for
the cases where we would be notified on a spend, act on in, but the
spending transaction did not get confirmed (e.g. a different transaction
would spend the same output). The logic for notifying such a spend is
not similar to the neutrino implementation, and will be unified at a
later point.