contractcourt/channel_arbitrator: only act on on-chain confirmation
This commit changes the channel arbitrator state machine to only care
about commitment transactions that are being confirmed on-chain. This
handles the cases where we would broadcast our commitment, expecting it
to get confirmed, but instead a competing transaction was confiremd.

The state machine has gotten a new state, StateCommitmentBroadcasted,
which we'll transition to after we have broadcasted our own commitment.
From this state we'll go to the StateContractClosed state regardless of
which commitment appears in the chain, unifying the contract resolution
betweee the local and remote force cloese.