channeldb/channel: define ClosureStatus, moved Borked to ChannelCloseSummary
This commit defines a new type `ClosureStatus` with the possible values
FullyClosed, PendingResolution, CommitmentBroadcasted, and Borked. This
is used in the ChannelCloseSummary as a replacement for IsPending, to
specify which kind of pending state the closing channel is in.

The ClosureType enum has gotten additional Unknown and LocalForcClosed.

The bool `IsBorked` in OpenChannel is deleted to be amonng the possible
ClosureStates. It will now be moved from the open channel bucket to the
closed channel bucket upon calling MarkBorked.

This means that OpenChannel now is used to _only_ specify actually open
channels, while all the (pending) close states are now part of the
ChannelCloseSummaries, through the different combinations of
ClosureState and ClosureType.