htlcswitch/link: make fail() call OnChannelFailure with LinkFailureError
This commit introduces a new error type LinkFailureError which is used
to distinguish the different kinds of errors that we can encounter
during link operation. It encapsulates the information necessary to
decide how we should handle the error.

In addition a new closure OnChannelFailure is added to the link config,
which is called when the link fails. This function closure should use
the given LinkFailureError to properly force close the channel, send an
error to the peer, and disconnect the peer.